GH-850 Login page redesigned
