--- front-end/src/layout/MinimalLayout/styles.module.css
@@ -1,37 +1,0 @@
-.MainWrapper {
-    background-color: #f5f5f5;
-}
-
-.Header {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    z-index: 1000;
-    width: 100%;
-    height: var(--axelix-header-height);
-    background-color: var(--axelix-primary-white);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    box-shadow: var(--axelix-box-shadow-light);
-}
-
-.LanguageSwitcherWrapper {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-}
-
-.ContentWrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: var(--axelix-header-height);
-    min-height: calc(100vh - var(--axelix-header-height));
-    padding-block: 30px;
-}
-
-.Logo {
-    width: 160px;
-}
