+++ front-end/src/layout/index.ts
@@ -15,5 +15,4 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-export { MinimalLayout } from "./MinimalLayout";
 export { MainLayout } from "./MainLayout";
