+++ front-end/src/pages/Login/LoginForm/styles.module.css
@@ -1,0 +1,24 @@
+.MainWrapper {
+    max-width: 400px;
+    width: 100%;
+    padding: 32px;
+    background-color: var(--axelix-primary-white);
+    border-radius: 16px;
+    box-shadow: var(--axelix-box-shadow-bold);
+}
+
+.LoginTitle {
+    text-align: center;
+    margin-bottom: 30px;
+}
+
+.SubmitButton {
+    height: 48px;
+    width: 100%;
+    margin-top: 16px;
+    font-weight: bold;
+}
+
+.ErrorAlert {
+    color: var(--axelix-primary-red);
+}
