+++ front-end/src/pages/Login/styles.module.css
@@ -1,28 +1,39 @@
-.LoginFormWrapper {
-    max-width: 400px;
+.MainWrapper {
+    display: flex;
+    min-height: 100vh;
     width: 100%;
-    padding: 32px;
+}
+
+.LeftImageWrapper {
+    height: 100vh;
+    width: 60%;
+}
+
+.Logo {
+    width: 150px;
+}
+
+.LeftImage {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    display: block;
+}
+
+.ContentWrapper {
+    width: 40%;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+    padding: 40px 40px 20px;
     background-color: var(--axelix-primary-white);
-    border-radius: 16px;
-    box-shadow: var(--axelix-box-shadow-bold);
 }
 
-.LoginTitle {
-    text-align: center;
-    margin-bottom: 30px;
-}
-
-.SubmitButton {
-    height: 48px;
+.ContentHeaderWrapper {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
     width: 100%;
-    margin-top: 16px;
-    font-weight: bold;
-}
-
-.ErrorAlert {
-    color: var(--axelix-primary-red);
-}
-
-.GreenLetter {
-    color: var(--axelix-primary-green);
+    column-gap: 20px;
 }
